rc5. Support panel 0.14 and use awesome-panel-cli
