--- mypy.ini
@@ -1,95 +1,0 @@
-[mypy]
-[mypy-sqlalchemy.orm]
-ignore_missing_imports=True
-[mypy-isodate]
-ignore_missing_imports=True
-[mypy-sqlalchemy]
-ignore_missing_imports=True
-[mypy-blpapi]
-ignore_missing_imports=True
-[mypy-setuptools.*]
-ignore_missing_imports=True
-[mypy-invoke.*]
-ignore_missing_imports=True
-[mypy-cx_Oracle.*]
-ignore_missing_imports = True
-[mypy-factory]
-ignore_missing_imports = True
-[mypy-freezegun]
-ignore_missing_imports = True
-[mypy-numpy]
-ignore_missing_imports = True
-[mypy-pandas.*]
-ignore_missing_imports = True
-[mypy-pytest]
-ignore_missing_imports = True
-[mypy-rest_framework.*]
-ignore_missing_imports = True
-[mypy-rest_pandas.*]
-ignore_missing_imports = True
-[mypy-django_filters.*]
-ignore_missing_imports = True
-[mypy-jsonfield.*]
-ignore_missing_imports = True
-[mypy-pyomo.*]
-ignore_missing_imports = True
-[mypy-pyutilib.*]
-ignore_missing_imports = True
-[mypy-networkx.*]
-ignore_missing_imports = True
-[mypy-matplotlib.*]
-ignore_missing_imports = True
-[mypy-sklearn.*]
-ignore_missing_imports = True
-[mypy-plotly.graph_objs.*]
-ignore_missing_imports = True
-[mypy-plotly.*]
-ignore_missing_imports = True
-[mypy-bokeh.*]
-ignore_missing_imports = True
-[mypy-sqlalchemy.*]
-ignore_missing_imports = True
-[mypy-winreg.*]
-ignore_missing_imports = True
-[mypy-selenium.*]
-ignore_missing_imports = True
-[mypy-holoviews.*]
-ignore_missing_imports = True
-[mypy-hvplot.*]
-ignore_missing_imports = True
-[mypy-panel.*]
-ignore_missing_imports = True
-[mypy-param.*]
-ignore_missing_imports = True
-[mypy-trading_analytics.utils.services.pandas_io.*]
-ignore_missing_imports = True
-[mypy-pyodbc.*]
-ignore_missing_imports = True
-[mypy-odlenv.*]
-ignore_missing_imports = True
-[mypy-odltransform.*]
-ignore_missing_imports = True
-[mypy-tqdm.*]
-ignore_missing_imports = True
-[mypy-awesome_panel_extensions.*]
-ignore_missing_imports = True
-[mypy-kubernetes.*]
-ignore_missing_imports = True
-[mypy-paramiko.*]
-ignore_missing_imports = True
-[mypy-dash.*]
-ignore_missing_imports = True
-[mypy-dash_bootstrap_components.*]
-ignore_missing_imports = True
-[mypy-dash_core_components.*]
-ignore_missing_imports = True
-[mypy-dash_table.*]
-ignore_missing_imports = True
-[mypy-dash_html_components.*]
-ignore_missing_imports = True
-[mypy-pypsa.*]
-ignore_missing_imports = True
-[mypy-entsoe.*]
-ignore_missing_imports = True
-[mypy-pyviz_comms.*]
-ignore_missing_imports = True
