--- panel_highcharts/.bokeh
@@ -1,1 +1,0 @@
-{"bokeh_version":"2.4.2","signatures":{"package.json":"84a84653b85ff9f202c74175ce58392b3bcdc176c7acce6c1a427dc532e4a893","package-lock.json":"4009183c7b3a33c9b7cb5dc8e5bbe33fd56203645c20edc4ac7dbad359273b6a","tsconfig.json":"87a823f0ef7885d0f8a99f845fc9b04922e05517668df9ba4dec17527eadeb41"}}
