--- pytest.ini
@@ -1,9 +1,0 @@
-[pytest]
-testpaths = tests
-junit_family=legacy
-python_files = tests.py test_*.py *_tests.py
-markers =
-    unittest: Small, isolated test
-    integration_test: A test towards some external system or service. For example a database
-    functionaltest: Large potentially non-isolated test requiring access to external services
-    slow: A slow test. Skip normally. But run this when you have the time.
