+++ src/panel_highcharts/.bokeh
@@ -1,0 +1,1 @@
+{"bokeh_version":"2.4.3","signatures":{"package.json":"c7678c10317da2d17d546b9db3415ccdaacd7d82dfa5750040b0632acdb2857d","package-lock.json":"a6adef216650fcd3434a3a131a9bce44e3fb2d2d2915f0587c01487365657cc0","tsconfig.json":"87a823f0ef7885d0f8a99f845fc9b04922e05517668df9ba4dec17527eadeb41"}}
