+++ src/panel_highcharts/__init__.py
@@ -5,3 +5,5 @@
 """
 from . import config
 from .charts import HighChart, HighGantt, HighMap, HighStock
+
+VERSION = "20221015.1rc1"
