+++ src/panel_highcharts/bokeh.ext.json
