+++ src/panel_highcharts/config.py
