+++ src/panel_highcharts/index.ts
