+++ src/panel_highcharts/models/highbase.py
