+++ src/panel_highcharts/models/highbase.ts
