+++ src/panel_highcharts/models/highchart.py
