+++ src/panel_highcharts/models/highgantt.ts
