+++ src/panel_highcharts/models/highmap.py
