+++ src/panel_highcharts/models/highmap.ts
