+++ src/panel_highcharts/models/highstock.py
