+++ src/panel_highcharts/models/highstock.ts
