+++ src/panel_highcharts/models/index.ts
