+++ src/panel_highcharts/tsconfig.json
